Time: 580 ms (14.40%) | Memory: 45.4 MB (54.80%) - LeetSync

+++ 2610-closest-prime-numbers-in-range/closest-prime-numbers-in-range.java
@@ -6,7 +6,7 @@ class Solution {
                 a.add(i);
             }
         }
-        double cnt=Integer.MAX_VALUE;int[] mn={-1,-1};
+        int cnt=Integer.MAX_VALUE;int[] mn={-1,-1};
         for(int j=0;j<a.size()-1;j++){
             if(a.get(j+1)-a.get(j)<cnt){
                 mn[0]=a.get(j);
